# backend/docs/swagger.yaml
basePath: /api/v1
definitions:
  dns.LookupRequest:
    properties:
      domain:
        type: string
    required:
    - domain
    type: object
  dns.LookupResponse:
    properties:
      domain:
        type: string
      error:
        type: string
      ipv4:
        items:
          type: string
        type: array
      ipv6:
        items:
          type: string
        type: array
      success:
        type: boolean
    type: object
info:
  contact: {}
  description: Backend API for the Mihombreng controller
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  title: Mihombreng API
  version: "1.0"
paths:
...
  /backup/create:
    post:
      description: Create a tar.gz backup of the entire working directory
      produces:
      - application/gzip
      responses:
        "200":
          description: Backup file
          schema:
            type: file
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Create backup
      tags:
      - Backup
  /backup/restore:
    post:
      consumes:
      - multipart/form-data
      description: Restore configuration from uploaded tar.gz backup file
      parameters:
      - description: Backup file (.tar.gz)
        in: formData
        name: backup
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties: true
            type: object
      summary: Restore backup
      tags:
      - Backup
  /dns/lookup:
    post:
      consumes:
      - application/json
      description: Resolve domain name to IPv4 and IPv6 addresses
      parameters:
      - description: Domain to lookup
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dns.LookupRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dns.LookupResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dns.LookupResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dns.LookupResponse'
      summary: Lookup domain IP addresses
      tags:
      - DNS
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"